# openapi.yaml
openapi: 3.0.1
info:
  title: Personal Budget
  version: 1.0.0
  description: >-
    A backend API to manage a personal budget using an envelope budgeting
    method.
  license:
    name: MIT
    url: https://choosealicense.com/licenses/mit/
paths:
  /api/v1/envelopes:
    summary: Gets and creates envelopes.
    description: >-
      This path is used to retrieve all of the envelopes from the database and
      create new envelopes.
    get:
      summary: Gets all the envelopes and their data.
      description: Retrieve the envelopes and their data from the database.
      operationId: get_envelopes
      responses:
        '200':
          description: Returns a list of all envelopes.
        '500':
          description: Internal Error.
      tags:
        - Envelopes
    post:
      summary: Creates a new envelope.
      description: >-
        Creates a new envelope by taking the title and budget of the envelope
        and inserting it into the database.
      operationId: add_envelope
      requestBody:
        description: Data for new envelope.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                budget:
                  type: number
              example:
                title: Scuba lessons
                budget: 300
      responses:
        '201':
          description: Envelope created successfully.
        '400':
          description: Body title or budget provided is not the proper type.
        '500':
          description: Internal Error.
      tags:
        - Add Envelope
  /api/v1/envelopes/{id}:
    summary: Performs operations on an envelope based on the provided id.
    description: >-
      This path is used to retrieve a specific envelope, update an envelope, and
      delete an envelope by using the provided id.
    get:
      summary: Gets an envelope with matching id.
      description: Retrieves an envelope that matches the provided id.
      operationId: get_envelope_by_id
      parameters:
        - name: id
          in: path
          description: The id of the envelope.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Returns the matching envelope.
        '404':
          description: Envelope with matching id not found.
        '500':
          description: Internal Error.
      tags:
        - Envelopes
    put:
      summary: Updates an envelope.
      description: Updates the matching envelope with the request body's title and budget.
      operationId: update_envelope
      parameters:
        - name: id
          in: path
          description: The id of the envelope.
          required: true
          schema:
            type: integer
      requestBody:
        description: Data to update the envelope.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                budget:
                  type: number
              example:
                title: Scuba lessons
                budget: 300
      responses:
        '201':
          description: Envelope updated successfully.
        '400':
          description: Body title or budget provided is not the proper type.
        '404':
          description: Envelope with matching id not found.
        '500':
          description: Internal Error.
      tags:
        - Update Envelope
    delete:
      summary: Deletes an envelope.
      description: Deletes the envelope with matching id from the database.
      operationId: delete_envelope
      parameters:
        - name: id
          in: path
          description: The id of the envelope.
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Envelope deleted successfully.
        '404':
          description: Envelope with matching id not found.
        '500':
          description: Internal Error.
      tags:
        - Delete Envelope
  /api/v1/envelopes/{fromId}/transfer/{toId}:
    summary: Transfers an amount from one envelope to another.
    description: >-
      This path is used to transfer a specific amount from the budget of one
      envelope to the budget of another.
    post:
      summary: Transfers an amount from one envelope to another.
      description: >-
        Transfers an amount from one envelope's budget to the other envelope's
        budget.
      operationId: transfer
      parameters:
        - name: fromId
          in: path
          description: The id of the envelope to transfer from.
          required: true
          schema:
            type: integer
        - name: toId
          in: path
          description: The id of the envelope to transfer to.
          required: true
          schema:
            type: integer
      requestBody:
        description: The amount to transfer.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
              example:
                amount: 200
      responses:
        '201':
          description: Amount transferred successfully.
        '400':
          description: >-
            Transfer amount provided is not a number or the amount to transfer
            exceeds the transferor envelope's budget.
        '404':
          description: One or more of the matching envelopes could not be found.
        '500':
          description: Internal Error.
      tags:
        - Transfer
  /api/v1/transactions:
    summary: Gets and creates transactions
    description: >-
      This path is used to retrieve all of the transactions in the database and
      create new transactions.
    get:
      summary: Gets all transactions
      description: Retrieves all transactions from the database and their associated data.
      operationId: get_transactions
      responses:
        '200':
          description: Returns a list of all transactions.
        '500':
          description: Internal Error.
      tags:
        - Transactions
    post:
      summary: Creates a new transactions
      description: >-
        Creates a new transaction, adds the transaction to the database, and
        updates the target envelope's budget.
      operationId: new_transactions
      requestBody:
        description: Data for new transaction.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                amount:
                  type: number
                recipient:
                  type: string
                envelopeId:
                  type: integer
              example:
                date: '2023-12-27'
                amount: 246.98
                recipient: Grocery Store
                envelopeId: 2
      responses:
        '201':
          description: Transaction created successfully.
        '400':
          description: >-
            Request body properties provided are not the proper types or the
            transaction amount exceeds the envelope's budget.
        '404':
          description: The provided envelope id does not match an existing envelope.
        '500':
          description: Internal Error.
      tags:
        - New Transaction
  /api/v1/transactions/{id}:
    summary: Performs operations on a transaction based on the provided id.
    description: >-
      This path is used to retrieve a specific transaction, update a
      transaction, and delete a transaction by using the provided id.
    get:
      summary: Gets the transaction with matching id.
      description: Retrieves a specific transaction from the database with a matching id.
      operationId: get_transaction_by_id
      parameters:
        - name: id
          in: path
          description: The id of the transaction.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Returns the matching transaction.
        '404':
          description: Transaction with matching id not found.
        '500':
          description: Internal Error.
      tags:
        - Transactions
    put:
      summary: Updates a transaction
      description: >-
        Updates the matching transaction with data provided in the body, and
        updates the target envelope's budget.
      operationId: update_transaction
      parameters:
        - name: id
          in: path
          description: The id of the transaction.
          required: true
          schema:
            type: integer
      requestBody:
        description: Data for new transaction.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                amount:
                  type: number
                recipient:
                  type: string
                envelopeId:
                  type: integer
              example:
                date: '2023-12-27'
                amount: 246.98
                recipient: Grocery Store
                envelopeId: 2
      responses:
        '201':
          description: Transaction updated successfully.
        '400':
          description: >-
            Request body properties provided are not the proper types or the
            transaction amount exceeds the envelope's budget.
        '404':
          description: Transaction or envelope with matching id not found.
        '500':
          description: Internal Error.
      tags:
        - Update Transaction
    delete:
      summary: Deletes a transaction.
      description: >-
        Removes a transaction from the database and adds the transaction amount
        back to the target envelope's budget.
      operationId: delete_transaction
      parameters:
        - name: id
          in: path
          description: The id of the transaction.
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Transaction deleted successfully.
        '404':
          description: Transaction with matching id not found.
        '500':
          description: Internal Error.
      tags:
        - Delete Transaction